chore: split graph
